+++ ubiconfig/_impl/loaders/base.py
@@ -8,13 +8,16 @@ class Loader(object):
     # This is a base class only. See implementations alongside this one
     # under the loaders module.
 
-    def load(self, file_name):
+    def load(self, file_name, version=None):
         """
         Load a single configuration file and return a :class:`UbiConfig` object.
 
         The given file_name should be a relative path to a YAML file
         in the loader's config source (e.g. relative path to file within a
         git repo or local directory).
+
+        version is an option argument to specify the version of config file to load,
+        if it's None, then use default, e.g. ubi7, ubi8
         """
         raise NotImplementedError()